Move IVsPerPropertyBrowsing to Interop
